Better TV emulation. Seems to fix problem in 50 Hz PAL games where frames go backwards.
git-svn-id: https://dolphin-emu.googlecode.com/svn/trunk@3811 8ced0084-cf51-0410-be5f-012b33b47a6e

+++ Source/Core/Core/Src/HW/VideoInterface.h
@@ -54,8 +54,8 @@ namespace VideoInterface
     void Update();
 
 	// urgh, ugly externs.
-	extern float ActualRefreshRate;
-	extern int TargetRefreshRate;
+	extern double ActualRefreshRate;
+	extern double TargetRefreshRate;
 	extern s64 SyncTicksProgress;
 
 	// UpdateInterrupts: check if we have to generate a new VI Interrupt
